# backend/swagger.yaml
openapi: 3.0.3
info:
  title: team-caltalk API
  version: 0.1.0
  description: "team-caltalk 개인 할일 관리 REST API. 근거 문서: docs/1-domain-definition.md v0.7, docs/2-PRD.md v0.9, docs/5-project-principle.md v0.7, docs/7-erd.md v0.1, docs/8-plan.md 3장(API 경로). JSON 필드는 camelCase, 날짜는 `YYYY-MM-DD` 문자열이다. 요청 본문은 100KB로 제한되며 초과 시 413이다 (NFR-10). `(가정)` 표시는 문서에 근거 없이 이 명세에서 정한 값이다."
servers:
  - url: /
    description: 프론트엔드와 같은 출처의 단일 서버 (PRD 7.2)
security:
  - bearerAuth: []
tags:
  - name: health
    description: 헬스체크
  - name: auth
    description: 회원가입·로그인·토큰 재발급·로그아웃 (PRD 6.4)
  - name: users
    description: 내 정보
  - name: categories
    description: 카테고리
  - name: todos
    description: 할일
paths:
  /api/health:
    get:
      tags:
        - health
      operationId: getHealth
      summary: 헬스체크
      description: DB에 `SELECT 1`을 실행하고 성공하면 200을 반환한다 (BE-02). 인증 불필요.
      security: []
      responses:
        "200":
          description: 정상
          content:
            application/json:
              schema:
                type: object
                required:
                  - status
                properties:
                  status:
                    type: string
                    enum:
                      - ok
                    description: 응답 형식 (가정)
  /api/auth/signup:
    post:
      tags:
        - auth
      operationId: signup
      summary: 회원가입
      description: FR-01, UC-01, BR-02, BR-05, BR-11. 사용자와 '기본' 카테고리를 하나의 트랜잭션으로 생성한다. 가입 후 토큰은 발급하지 않고 화면은 로그인으로 이동한다 (도메인 7.1). 중복 이메일은 400이며 사유는 `email` 항목에 달린다. 인증 불필요.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "201":
          description: 가입 완료. 생성된 사용자를 반환한다 (상태 코드·응답 본문 가정)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/auth/login:
    post:
      tags:
        - auth
      operationId: login
      summary: 로그인
      description: FR-02, UC-02, BR-14, PRD 6.4. Access Token은 본문, Refresh Token은 `Set-Cookie`로 발급한다. 이메일 없음·비밀번호 틀림을 구분하지 않고 같은 401과 같은 문구로 응답한다 (BR-14). 인증 불필요.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          $ref: "#/components/responses/TokenIssued"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: 인증 실패. 실패 사유를 구분하지 않는다 (BR-14)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                message: 식별자 또는 비밀번호가 올바르지 않습니다
        "429":
          description: 같은 IP의 로그인 실패가 분당 10회를 초과해 일시 거부 (NFR-11, P1). 제한 중에는 올바른 비밀번호 요청도 1분간 거부한다 (가정). 상태 코드는 가정
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/refresh:
    post:
      tags:
        - auth
      operationId: refreshToken
      summary: 토큰 재발급
      description: FR-04, PRD 6.4. Refresh 쿠키의 서명·만료·`type`을 검증하고 DB의 `jti`를 확인한 뒤 기존 행 삭제 + 새 Refresh 저장(회전)과 새 Access를 발급한다. 서명은 유효하지만 `jti`가 DB에 없으면 재사용으로 보고 해당 사용자의 Refresh를 모두 삭제하고 401을 반환한다.
      security:
        - refreshCookie: []
      responses:
        "200":
          $ref: "#/components/responses/TokenIssued"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/auth/logout:
    post:
      tags:
        - auth
      operationId: logout
      summary: 로그아웃
      description: FR-03, UC-11, PRD 6.4. 쿠키의 `jti` 행을 삭제하고 Refresh 쿠키를 삭제한다. 클라이언트는 메모리의 Access Token을 폐기한다. 쿠키가 없거나 무효여도 204로 응답한다 (가정).
      security:
        - refreshCookie: []
      responses:
        "204":
          description: 로그아웃 완료 (상태 코드 가정)
          headers:
            Set-Cookie:
              description: "Refresh 쿠키 삭제. 예: `refreshToken=; Path=/api/auth; Max-Age=0; HttpOnly; SameSite=Strict`"
              schema:
                type: string
  /api/users/me:
    get:
      tags:
        - users
      operationId: getMe
      summary: 내 정보 조회
      description: 헤더의 이름 표시용 (8-plan 3장, 가정). Access Token의 `sub` 사용자 정보를 반환한다 (BR-04).
      responses:
        "200":
          description: 내 정보
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
    patch:
      tags:
        - users
      operationId: updateMe
      summary: 이름 변경
      description: FR-13(P1), UC-03, BR-04, BR-16. 이름만 변경한다. 이메일은 변경할 수 없으며 본문의 `email` 필드는 무시한다 (가정).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: 변경된 내 정보 (응답 본문 가정)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/users/me/password:
    put:
      tags:
        - users
      operationId: changePassword
      summary: 비밀번호 변경
      description: FR-13(P1), UC-03, BR-15, PRD 6.4. 현재 비밀번호를 확인하고, 불일치하면 400(`currentPassword` 항목 사유)으로 거부한다. 성공 시 해당 사용자의 Refresh를 모두 삭제해 다른 기기를 로그아웃시키고, 현재 기기에는 새 Access(본문)와 새 Refresh(`Set-Cookie`)를 발급한다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordChangeRequest"
      responses:
        "200":
          $ref: "#/components/responses/TokenIssued"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/categories:
    get:
      tags:
        - categories
      operationId: listCategories
      summary: 카테고리 목록 조회
      description: FR-05, UC-12, BR-03. 본인 카테고리만 반환하며 '기본'이 포함된다. 정렬 순서는 정하지 않는다 (가정).
      responses:
        "200":
          description: 본인 카테고리 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags:
        - categories
      operationId: createCategory
      summary: 카테고리 생성
      description: FR-06, UC-04, 도메인 3.2. 소유자 내 중복 이름은 400(`name` 항목 사유)이다. 다른 사용자와 같은 이름은 허용한다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: 생성된 카테고리 (상태 코드 가정)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/categories/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    patch:
      tags:
        - categories
      operationId: renameCategory
      summary: 카테고리 이름 변경
      description: FR-16, UC-13, BR-17, AC-10-1, AC-10-2, AC-10-4. 소유자 내 중복 이름은 400이다. '기본' 카테고리 변경 요청은 400(`name` 항목 사유)으로 거부한다 (상태 코드 가정). 본인 카테고리가 아니거나 없으면 404 (BR-03).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "200":
          description: 변경된 카테고리
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags:
        - categories
      operationId: deleteCategory
      summary: 카테고리 삭제
      description: FR-17, UC-14, BR-17, BR-18, AC-10-3, AC-10-4. 소속 할일을 본인 '기본' 카테고리로 옮긴 뒤 카테고리를 삭제한다 (하나의 트랜잭션). 할일은 삭제하지 않는다. '기본' 카테고리 삭제 요청은 400(`name` 항목 사유)으로 거부한다 (상태 코드 가정). 본인 카테고리가 아니거나 없으면 404 (BR-03).
      responses:
        "204":
          description: 삭제 완료 (상태 코드 가정)
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/todos:
    get:
      tags:
        - todos
      operationId: listTodos
      summary: 할일 목록 조회 및 필터
      description: FR-10, UC-08, BR-10, AC-08, OI-05, OI-09. 본인 할일만 반환한다 (BR-03). 상태는 `today` 기준으로 도출해 응답에 포함한다. 필터는 단일 선택이며 `categoryId`를 주면 `filter`는 `all`이어야 한다 (가정, 위반 시 400, `filter` 항목 사유). 정렬은 종료일자 오름차순, 같으면 등록순이다.
      parameters:
        - name: filter
          in: query
          required: false
          description: 상태 필터. `all`=전체, `not_started`=시작 전, `in_progress`=진행 중, `done`=완료, `overdue`=지연
          schema:
            type: string
            enum:
              - all
              - not_started
              - in_progress
              - done
              - overdue
            default: all
        - name: categoryId
          in: query
          required: false
          description: 카테고리별 필터. 본인 카테고리가 아니면 빈 목록을 반환한다 (가정, BR-03)
          schema:
            type: integer
            format: int64
        - $ref: "#/components/parameters/Today"
      responses:
        "200":
          description: 할일 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Todo"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags:
        - todos
      operationId: createTodo
      summary: 할일 등록
      description: FR-07, UC-05, BR-06, BR-07, BR-08, BR-11, BR-13, AC-05. `categoryId`가 없거나 null이면 본인 '기본' 카테고리를 적용한다. 다른 사용자의 카테고리를 지정하면 400(`categoryId` 항목 사유)으로 거부한다 (BR-13, 상태 코드 가정). 종료일자 < 시작일자는 400(`endDate` 항목 사유).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TodoCreate"
      responses:
        "201":
          description: 등록된 할일. 응답에는 `status`가 없다 (등록·수정 요청에 `today`가 없으므로, 가정)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TodoBase"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/todos/calendar:
    get:
      tags:
        - todos
      operationId: listCalendarTodos
      summary: 월별 캘린더 할일 조회
      description: FR-11, UC-09, BR-10, BR-12, AC-09, OI-05. `시작일자 <= 월 말일 AND 종료일자 >= 월 1일`인 본인 할일을 반환한다. 완료된 할일도 포함한다 (OI-10). 정렬은 목록과 같다 (가정).
      parameters:
        - name: month
          in: query
          required: true
          description: 조회할 월 (YYYY-MM)
          schema:
            type: string
            pattern: ^\d{4}-\d{2}$
          example: 2026-10
        - $ref: "#/components/parameters/Today"
      responses:
        "200":
          description: 해당 월에 걸친 할일 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Todo"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/todos/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    patch:
      tags:
        - todos
      operationId: updateTodo
      summary: 할일 수정 / 완료 토글
      description: FR-08, FR-14, UC-06, BR-06, BR-08, BR-11, BR-13, AC-06. 보낸 필드만 바꾼다. 수정할 필드가 없는 본문은 400(`body` 항목 사유, 가정). `categoryId`를 null로 보내면 본인 '기본' 카테고리를 적용한다 (BR-06). 수정 후 종료일자 < 시작일자가 되면 400이고 저장된 값은 바뀌지 않는다 (AC-06-3). 본인 할일이 아니거나 없으면 404 (BR-03, AC-06-4).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TodoUpdate"
      responses:
        "200":
          description: 수정된 할일 (`status` 없음, 가정)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TodoBase"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags:
        - todos
      operationId: deleteTodo
      summary: 할일 삭제
      description: FR-09, UC-07, AC-07, OI-11. 영구 삭제한다. 본인 할일이 아니거나 없으면 404 (BR-03, AC-07-3).
      responses:
        "204":
          description: 삭제 완료 (상태 코드 가정)
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Access Token. JWT HS256, payload `sub`, `type: "access"`, 만료 15분 (PRD 6.4). `Authorization: Bearer <token>`'
    refreshCookie:
      type: apiKey
      in: cookie
      name: refreshToken
      description: 'Refresh Token 쿠키. 쿠키 이름 `refreshToken`은 가정. JWT HS256, payload `sub`, `jti`, `type: "refresh"`, 만료 7일. `HttpOnly; SameSite=Strict; Path=/api/auth` (운영 `Secure`). `/api/auth/refresh`, `/api/auth/logout`에만 전송된다 (PRD 6.4)'
  parameters:
    Id:
      name: id
      in: path
      required: true
      description: 리소스 ID
      schema:
        type: integer
        format: int64
    Today:
      name: today
      in: query
      required: true
      description: 사용자 로컬 기준 오늘 날짜 (YYYY-MM-DD). 서버는 형식만 검증하고 상태 판정에 사용한다 (OI-05, PRD 9장)
      schema:
        type: string
        format: date
      example: "2026-10-01"
  headers:
    SetRefreshCookie:
      description: "Refresh Token 발급. 예: `refreshToken=<JWT>; HttpOnly; SameSite=Strict; Path=/api/auth` (운영 `Secure`). 쿠키 이름은 가정"
      schema:
        type: string
  responses:
    TokenIssued:
      description: 토큰 발급. Access는 본문, Refresh는 `Set-Cookie` (PRD 6.4)
      headers:
        Set-Cookie:
          $ref: "#/components/headers/SetRefreshCookie"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AccessTokenResponse"
    BadRequest:
      description: 입력 검증 실패. 위반 항목별 사유를 포함한다 (BR-11, NFR-10)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            message: 입력값을 확인하세요
            errors:
              - field: endDate
                reason: 종료일자는 시작일자보다 빠를 수 없습니다
    Unauthorized:
      description: 인증 실패. 토큰 없음·만료·위조 (BR-01, PRD 6.4)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: 찾을 수 없음. 본인 소유가 아니거나 존재하지 않는다 (BR-03, NFR-09)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: 사용자에게 보여줄 오류 문구
        errors:
          type: array
          description: 400일 때 위반 항목별 사유 (BR-11)
          items:
            type: object
            required:
              - field
              - reason
            properties:
              field:
                type: string
                description: 요청 필드 이름 (camelCase)
              reason:
                type: string
                description: 위반 사유
    AccessTokenResponse:
      type: object
      required:
        - accessToken
      properties:
        accessToken:
          type: string
          description: Access Token (JWT). 클라이언트는 메모리에만 보관한다 (PRD 6.4)
    User:
      type: object
      required:
        - id
        - email
        - name
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
          format: email
          maxLength: 254
          description: 로그인 식별자. 변경 불가 (BR-16)
        name:
          type: string
          minLength: 1
          maxLength: 30
    SignupRequest:
      type: object
      required:
        - email
        - password
        - name
      properties:
        email:
          type: string
          format: email
          maxLength: 254
          description: 앞뒤 공백만 제거하고 대소문자는 구분한다 (가정, 7-erd 4장 5)
        password:
          $ref: "#/components/schemas/Password"
        name:
          type: string
          minLength: 1
          maxLength: 30
          description: 앞뒤 공백 제거 후 1~30자
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
        password:
          type: string
    UserUpdate:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 30
          description: 앞뒤 공백 제거 후 1~30자
    PasswordChangeRequest:
      type: object
      required:
        - currentPassword
        - newPassword
      properties:
        currentPassword:
          type: string
          description: 현재 비밀번호 (BR-15)
        newPassword:
          $ref: "#/components/schemas/Password"
    Password:
      type: string
      minLength: 8
      maxLength: 64
      pattern: ^(?=.*[A-Za-z])(?=.*\d).+$
      description: 8~64자, 영문·숫자 각 1자 이상 (OI-02 가정). 공백 제거 없이 입력 그대로 센다
    Category:
      type: object
      required:
        - id
        - name
        - isDefault
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          minLength: 1
          maxLength: 20
        isDefault:
          type: boolean
          description: "'기본' 카테고리 여부. 사용자당 1개 (BR-05)"
    CategoryInput:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 20
          description: 앞뒤 공백 제거 후 1~20자, 소유자 내 고유
    TodoBase:
      type: object
      required:
        - id
        - categoryId
        - categoryName
        - title
        - description
        - startDate
        - endDate
        - isCompleted
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
          format: int64
        categoryId:
          type: integer
          format: int64
        categoryName:
          type: string
          description: 목록·캘린더 표시용 카테고리 이름 (가정)
        title:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 1000
          nullable: true
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        isCompleted:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Todo:
      description: 조회 응답용 할일. `status`는 요청의 `today` 기준으로 도출한다 (BR-10)
      allOf:
        - $ref: "#/components/schemas/TodoBase"
        - type: object
          required:
            - status
          properties:
            status:
              $ref: "#/components/schemas/TodoStatus"
    TodoStatus:
      type: string
      enum:
        - not_started
        - in_progress
        - done
        - overdue
      description: "`not_started`=시작 전, `in_progress`=진행 중, `done`=완료, `overdue`=지연 (BR-10)"
    TodoCreate:
      type: object
      required:
        - title
        - startDate
        - endDate
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 100
          description: 앞뒤 공백 제거 후 1~100자
        description:
          type: string
          maxLength: 1000
          nullable: true
          description: 선택. 빈 값은 null로 저장
        categoryId:
          type: integer
          format: int64
          nullable: true
          description: 없거나 null이면 '기본' 카테고리 (BR-06)
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
          description: 시작일자 이상 (BR-08)
    TodoUpdate:
      type: object
      description: 보낸 필드만 수정한다
      minProperties: 1
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 100
          description: 앞뒤 공백 제거 후 1~100자
        description:
          type: string
          maxLength: 1000
          nullable: true
          description: 빈 값 또는 null이면 null로 저장
        categoryId:
          type: integer
          format: int64
          nullable: true
          description: null이면 '기본' 카테고리 (BR-06)
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
          description: 시작일자 이상 (BR-08)
        isCompleted:
          type: boolean
          description: 완료 / 완료 해제 (FR-08, FR-14)